backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
pydantic==2.6.0
pydantic-extra-types==2.6.0
pydantic-settings==2.2.0
python-dotenv==1.0.1
email-validator==2.1.0
aiosqlite==0.20.0
pytest==8.0.0
pytest-asyncio==0.23.5
httpx==0.27.0
python-multipart==0.0.28
